feat: Add getUser function to retrieve user information
This commit adds a new function, `getUser`, to the `AuthAPI.ts` file. The function makes a GET request to the `/auth/user` endpoint to retrieve the user information. The response is then parsed using the `userSchema` and returned if successful. If there is an error, an appropriate error message is thrown.

Recent user commits:
- feat: Add request interceptor to include authorization token
- feat: Update user schema and types
- feat: Add user endpoint to retrieve user information, to authenticate and autorizate