feat: Add manager field to project schema
This commit adds the `manager` field to the `projectSchema` in the `index.ts` file. The `manager` field is of type `string` and is required. It is used to store the ID of the project manager. This change allows for better organization and management of projects by associating them with specific managers.