refactor: Update TaskController to populate completedBy field in getTasksById
This commit updates the `TaskController` to populate the `completedBy` field in the `getTasksById` method. The `completedBy` field is populated with the `id`, `name`, and `email` of the user who completed the task. This change allows for better retrieval of task details, including the information of the user who completed the task.