refactor: Update `Note` model schema to include createdBy field
This commit updates the `Note` model schema in the `Note.ts` file to include the `createdBy` field. The `createdBy` field is of type `Types.ObjectId` and is required. It references the `User` model and represents the user who created the note. This change improves the data structure and allows for better tracking of note creators.