refactor: Add checkPassword endpoint for validating user password
This commit adds the checkPassword endpoint to the authRoutes.ts file. The endpoint allows users to validate their password by comparing it with the stored password in the database. If the password is valid, a success response is returned. If the password is invalid, an error response is returned. This addition enhances the security of the application by providing a mechanism to verify user passwords.